Update TypeOfCarePage to use redux hooks
